% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_index.R
\name{extract_batch_ima}
\alias{extract_batch_ima}
\title{extract_batch_ima}
\usage{
extract_batch_ima(ini_date, end_date)
}
\arguments{
\item{end_date}{}
}
\value{

}
\description{
extract_batch_ima
}
\examples{
extract_daily_ima("05/05/2020")
}
