% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_index.R
\name{extract_batch_usdbrl}
\alias{extract_batch_usdbrl}
\title{Dolar}
\usage{
extract_batch_usdbrl(ini_date, end_date)
}
\arguments{
\item{end_date}{}
}
\value{

}
\description{
https://br.investing.com/currencies/usd-brl-historical-data
}
