% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fund.R
\name{extract_funds_by_cnpj_list}
\alias{extract_funds_by_cnpj_list}
\title{extract_funds_by_cnpj_list}
\usage{
extract_funds_by_cnpj_list(ini_date, end_date, cnpj_list)
}
\arguments{
\item{cnpj_list}{}
}
\value{

}
\description{
Extrai o valor de cota de fundos listados em um intervalo de tempo
}
