% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fund.R
\name{get_funds_at_least}
\alias{get_funds_at_least}
\title{get_funds_at_least}
\usage{
get_funds_at_least(min_cotistas, min_pl)
}
\arguments{
\item{min_pl}{}
}
\value{

}
\description{
Gera uma lista com os fundos que tenham um número mínimo de cotistas e de PL.
}
