% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_index.R
\name{spawn_index_dataframe}
\alias{spawn_index_dataframe}
\title{spawn_index_dataframe}
\usage{
spawn_index_dataframe(ini_date, end_date)
}
\arguments{
\item{end_date}{}
}
\value{

}
\description{
spawn_index_dataframe
}
